pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lightspeed-agent"
version = "0.1.0"
description = "A2A-ready agent for Red Hat Insights using Google ADK"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
authors = [
    { name = "Red Hat", email = "insights@redhat.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "google-adk>=0.5.0",
    "a2a-sdk[http-server]>=0.3.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "uvicorn>=0.30.0",
    "fastapi>=0.115.0",
    "PyJWT[crypto]>=2.8.0",
    "cryptography>=42.0.0",
    "redis>=5.0.0",
    "google-cloud-service-control>=1.0.0",
    "google-cloud-pubsub>=2.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "aiosqlite>=0.20.0",
    # PostgreSQL drivers
    "asyncpg>=0.29.0",  # Async driver for SQLAlchemy async operations
    "psycopg2-binary>=2.9.0",  # Sync driver for ADK DatabaseSessionService
    # OpenTelemetry for distributed tracing
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-httpx>=0.41b0",
    "opentelemetry-instrumentation-logging>=0.41b0",
]

[project.optional-dependencies]
telemetry = [
    "opentelemetry-exporter-jaeger>=1.20.0",
    "opentelemetry-exporter-zipkin>=1.20.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "httpx>=0.27.0",
]

[project.scripts]
lightspeed-agent = "lightspeed_agent.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/lightspeed_agent"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
